+++ crates/lynx-proxy/src/routes/network/components/Contents/TextViewer/index.tsx
@@ -2,20 +2,22 @@ import React, { useMemo } from 'react';
 
 interface TextViewProps {
   arrayBuffer?: ArrayBuffer;
+  text?: string;
 }
 
-const TextView: React.FC<TextViewProps> = ({ arrayBuffer }) => {
+const TextView: React.FC<TextViewProps> = ({ arrayBuffer, text }) => {
   const data = useMemo(() => {
+    if (text) return text;
     if (!arrayBuffer) return null;
 
-    return new TextDecoder('utf-8').decode(new Uint8Array(arrayBuffer));
-  }, [arrayBuffer]);
+    return new TextDecoder().decode(new Uint8Array(arrayBuffer));
+  }, [arrayBuffer, text]);
 
   if (!data) return null;
 
   return (
     <div className="flex h-full flex-col rounded-sm border border-gray-300 bg-white p-1 font-mono text-xs text-black dark:border-zinc-700 dark:bg-zinc-900 dark:text-zinc-200">
-      <pre>{data}</pre>
+      <pre className="whitespace-normal break-all">{data}</pre>
     </div>
   );
 };