+++ crates/lynx-proxy/src/routes/network/components/RecordingStatusButton/index.tsx
@@ -1,9 +1,11 @@
 import {
+  getGetCaptureStatusQueryKey,
   useGetCaptureStatus,
   useToggleCapture,
 } from '@/services/generated/net-request/net-request';
 import { RecordingStatus } from '@/services/generated/utoipaAxum.schemas';
 import { RiPauseLargeFill, RiPlayLargeFill } from '@remixicon/react';
+import { useQueryClient } from '@tanstack/react-query';
 import { Button } from 'antd';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
@@ -14,15 +16,24 @@ export const RecordingStatusButton: React.FC<
   IRecordingStatusButtonProps
 > = () => {
   const { t } = useTranslation();
-  const { data: netWorkCaptureStatusData, refetch } = useGetCaptureStatus();
-  const toggleCapture = useToggleCapture();
+  const queryClient = useQueryClient();
+  const { data: netWorkCaptureStatusData } = useGetCaptureStatus();
+  const toggleCapture = useToggleCapture({
+    mutation: {
+      onSuccess: () => {
+        queryClient.invalidateQueries({
+          queryKey: getGetCaptureStatusQueryKey(),
+          refetchType: 'active',
+        });
+      },
+    },
+  });
   const recordingStatus = netWorkCaptureStatusData?.data.recordingStatus;
   return (
     <Button
       type="text"
       onClick={async () => {
         await toggleCapture.mutateAsync();
-        await refetch();
       }}
       icon={
         recordingStatus === RecordingStatus.startRecording ? (